未使用引数のコメントアウトと複数のバグ修正
- 多数のメソッドで未使用引数をコメントアウト
- `Camera::createOrthographic` メソッドの引数変更
- `AudioCache` のメンバ変数 `m_alBufferId` を `ALuint` 型にキャスト
- `openWav` 関数から `filename` 引数を削除
- `AudioEngine` クラスに `m_audioPathIdMap` を追加
- `AudioEngineImpl::uncache` メソッドのスペルミス修正
- `Console.cpp` の `vformat` 関数から未使用変数 `ret` を削除
- `EventListener` と `EventMouse` クラスのコンストラクタでメンバ変数を初期化
- `FileUtils::addSearchPath` メソッドに重複パス追加防止ロジックを追加
- `Keyboard` クラスのメソッドで型キャストと文字列フォーマットを修正
- `OpenGLUtility::getGLDataTypeSize` メソッドに `break` ステートメントを追加
- `Program.cpp` の `setUniform` 関数で `count` を使用
- `Renderer` クラスのメソッドでインデックス計算を修正
- `ShaderManager::loadProgram` メソッドが `Program` オブジェクトを返すように修正
- `TextureAtlas::setupIndices` メソッドでインデックスを `unsigned short` 型にキャスト

+++ src/renderer/TextureAtlas.cpp
@@ -116,13 +116,13 @@ void TextureAtlas::setupIndices()
     if (m_capacity == 0) return;
 
     for (int i = 0; i < m_capacity; i++) {
-        m_pIndices[i * 6 + 0] = i * 4 + 0;
-        m_pIndices[i * 6 + 1] = i * 4 + 1;
-        m_pIndices[i * 6 + 2] = i * 4 + 2;
+        m_pIndices[i * 6 + 0] = static_cast<unsigned short>(i * 4 + 0);
+        m_pIndices[i * 6 + 1] = static_cast<unsigned short>(i * 4 + 1);
+        m_pIndices[i * 6 + 2] = static_cast<unsigned short>(i * 4 + 2);
 
-        m_pIndices[i * 6 + 3] = i * 4 + 3;
-        m_pIndices[i * 6 + 4] = i * 4 + 2;
-        m_pIndices[i * 6 + 5] = i * 4 + 1;
+        m_pIndices[i * 6 + 3] = static_cast<unsigned short>(i * 4 + 3);
+        m_pIndices[i * 6 + 4] = static_cast<unsigned short>(i * 4 + 2);
+        m_pIndices[i * 6 + 5] = static_cast<unsigned short>(i * 4 + 1);
     }
 }
